30% Group Project/Group A1/Server_Py/requirements.txt
mysql-connector-python==8.0.33
